+++ mayan/apps/smart_settings/views/setting_views.py
@@ -1,70 +1,18 @@
 from django.conf import settings
 from django.contrib import messages
-from django.http import Http404
-from django.urls import reverse, reverse_lazy
+from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
 from mayan.apps.views.generics import (
-    ConfirmView, FormView, MultipleObjectConfirmActionView,
-    SingleObjectListView
+    FormView, MultipleObjectConfirmActionView
 )
 
-from .forms import SettingForm
-from .icons import (
-    icon_setting_cluster_configuration_save,
-    icon_setting_cluster_namespace_list, icon_setting_namespace_detail,
-    icon_setting_edit, icon_setting_revert
-)
-from .literals import MESSAGE_LOCAL_STORAGE_DISABLED
-from .permissions import permission_settings_edit, permission_settings_view
-from .setting_domains.models import SettingDomainModel
-from .settings import setting_cluster
-
-
-class SettingClusterConfigurationFileSave(ConfirmView):
-    post_action_redirect = reverse_lazy(
-        viewname='settings:setting_cluster_namespace_list'
-    )
-    view_icon = icon_setting_cluster_configuration_save
-    view_permission = permission_settings_edit
-
-    def dispatch(self, request, *args, **kwargs):
-        if settings.COMMON_DISABLE_LOCAL_STORAGE:
-            messages.warning(
-                message=MESSAGE_LOCAL_STORAGE_DISABLED, request=self.request
-            )
-
-        return super().dispatch(request=request, *args, **kwargs)
-
-    def get_extra_context(self):
-        return {
-            'message': _(
-                message='This will overwrite the content of the '
-                'configuration file.'
-            ),
-            'title': _(message='Save settings to the configuration file?')
-        }
-
-    def view_action(self, form=None):
-        setting_cluster.do_make_persistent()
-
-        messages.success(
-            message=_(
-                message='Settings saved to configuration file successfully.'
-            ), request=self.request
-        )
-
-
-class SettingClusterNamespaceListView(SingleObjectListView):
-    extra_context = {
-        'hide_link': True,
-        'title': _(message='Setting namespaces')
-    }
-    view_icon = icon_setting_cluster_namespace_list
-    view_permission = permission_settings_view
-
-    def get_source_queryset(self):
-        return setting_cluster.get_namespace_list()
+from ..forms import SettingForm
+from ..icons import icon_setting_edit, icon_setting_revert
+from ..literals import MESSAGE_LOCAL_STORAGE_DISABLED
+from ..permissions import permission_settings_edit
+from ..setting_domains.models import SettingDomainModel
+from ..settings import setting_cluster
 
 
 class SettingValueEditView(FormView):
@@ -165,37 +113,3 @@ class SettingValueRevertView(MultipleObjectConfirmActionView):
 
     def object_action(self, form, instance):
         instance.do_value_revert()
-
-
-class SettingNamespaceDetailView(SingleObjectListView):
-    view_icon = icon_setting_namespace_detail
-    view_permission = permission_settings_view
-
-    def get_extra_context(self):
-        namespace = self.get_namespace()
-
-        return {
-            'hide_object': True,
-            'object': namespace,
-            'subtitle': _(
-                message='Settings inherited from an environment variable '
-                'take precedence and cannot be changed in this view. '
-            ),
-            'title': _(message='Settings in namespace: %s') % namespace
-        }
-
-    def get_namespace(self):
-        try:
-            return setting_cluster.get_namespace(
-                name=self.kwargs['namespace_name']
-            )
-        except KeyError:
-            raise Http404(
-                _(message='Namespace: %s, not found') % self.kwargs[
-                    'namespace_name'
-                ]
-            )
-
-    def get_source_queryset(self):
-        setting_namespace = self.get_namespace()
-        return setting_namespace.get_setting_list()