+++ HISTORY.rst
@@ -175,6 +175,8 @@
 
   - Fancybox from 3.2.5 to 3.3.5
 
+- Split smart settings app views module.
+
 4.10.3 (2025-12-24)
 ===================
 - Include changes and fixes from versions 4.9.8 and 4.9.7.